backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.0
xgboost>=1.7.0
python-dotenv==1.0.0
matplotlib==3.7.0
requests==2.31.0
gunicorn==21.2.0
